Migrate onLogin to Login component
This functionality still needs to be refactored, however, it does not
need to live in a separate location from the login form.
